🤖 avoid async timeout test conflicts
