+++ osprey_async_worker/src/osprey/async_worker/stdlib_udfs/async_has_label.py
@@ -1,5 +1,3 @@
-# pattern: Imperative Shell
-
 """native async `HasLabel`"""
 
 from typing import Any, Sequence, cast
